humility-core: core: change register interface to use `u64` over `u32`.
Note that this does not change how registers are displayed or handled,
it only changes the interface for interacting with the core.

+++ cmd/tasks/src/lib.rs
@@ -206,7 +206,8 @@ fn tasks(context: &mut humility::ExecutionContext) -> Result<()> {
         core.halt()?;
 
         let cur =
-            hubris.arch.as_ref().unwrap().get_current_task_ptr(hubris, core)?;
+            hubris.arch.as_ref().unwrap().get_current_task_ptr(hubris, core)?
+                as u32;
 
         //
         // We read the entire task table at a go to get as consistent a
@@ -258,7 +259,7 @@ fn tasks(context: &mut humility::ExecutionContext) -> Result<()> {
         }
 
         if let Ok(pc) = core.read_reg(hubris.arch.as_ref().unwrap().get_pc()) {
-            if hubris.instr_mod(pc).is_none() {
+            if hubris.instr_mod(pc as u32).is_none() {
                 humility::warn!(
                     "PC 0x{:x} is unknown; \
                     system may be executing in ROM!",
